midi stuff

+++ code/ToolKit/DxlBehav15/Sensors.pde
@@ -6,15 +6,15 @@ class SensorEvt
   float   value  = 0;
   float   center = 0;
   float   coef   = 1.0f;
-  int     min = -1024;
-  int     max =  1024;
+  int     min    = -1024;
+  int     max    =  1024;
   String  cmd = null;
   void    fromXML(XML xml)
   {
     servo  = xml.getInt("servo");
-    value  = xml.getInt("value");
-    center = xml.getInt("center");
-    coef   = xml.getInt("coef");
+    value  = xml.getFloat("value");
+    center = xml.getFloat("center");
+    coef   = xml.getFloat("coef");
     min    = xml.getInt("min");
     max    = xml.getInt("max");
     cmd    = xml.getString("cmd");
